third_party/blink/web_tests/external/wpt/referrer-policy/generic/inheritance/iframe-inheritance-javascript-expected.txt

This is a testharness.js-based test.
PASS Referrer Policy: iframes with javascript url reuse referrer policy
FAIL Referrer Policy: iframes with javascript url reuse referrer policy 1 assert_equals: expected "http://web-platform.test:8001/" but got "http://web-platform.test:8001/custom"
Harness: the test ran to completion.
